move stat formatting to separate file
