+++ src/gamemodes/versus/mod.rs
@@ -1,0 +1,2 @@
+pub mod versus;
+pub mod garbage;
