+++ src/ui_components/mod.rs
@@ -7,3 +7,5 @@ pub mod start_screen;
 pub mod versus_ready;
 pub mod high_score;
 pub mod input_name;
+
+pub mod stat_formatting;
